[PANA-8393] Forward embedded content resources
